Hit me baby one more time.
We fixed the GUID in the previous commit, but didn't pay attention that the sources in the FMU were still the b0rked 1.25.x ones!

Re-roll, with a fresh Linux `.so`, compiled from 1.26-nightly, AND include actual source.
Correspondingly, this means that the Windows DLL is mildly out-of-sync with the contained sources, but this is a sacrifice I'm willing to make.